requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
pymongo==4.4.1
requests==2.31.0
Pillow==10.0.0
python-dotenv==1.0.0
google-generativeai==0.8.4
vertexai==1.67.0
openai==1.57.4
